This commit applies the component composition pattern to all domain screens.
- All domain screens (`user`, `learning`, `engagement`, `analytics`) have been refactored to be composed of components from their respective interfaces.
- The UI from each domain screen has been extracted into new, more granular components within the `components` directory of each domain.
- The domain interfaces have been updated to expose new render methods for these components, following the established architectural pattern.
