+++ client/src/engagement/components/AchievementsView.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View, Text } from 'react-native-ui-lib';
+
+const AchievementsView = () => {
+  return (
+    <View flex center>
+      <Text text50b marginB-20>Achievements</Text>
+      <Text>(A list of user achievements would be displayed here.)</Text>
+    </View>
+  );
+};
+
+export default AchievementsView;
