+++ client/src/engagement/components/LeaderboardView.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View, Text } from 'react-native-ui-lib';
+
+const LeaderboardView = () => {
+  return (
+    <View flex center>
+      <Text text50b marginB-20>Leaderboard</Text>
+      <Text>(The leaderboard would be displayed here.)</Text>
+    </View>
+  );
+};
+
+export default LeaderboardView;
