+++ client/src/engagement/screens/LeaderboardScreen.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
-import { View, Text } from 'react-native-ui-lib';
+import { EngagementComponent } from '..';
 import { withEngagementData } from '../../common/hocs/withInitialData';
 
+const engagement = new EngagementComponent();
+
 const LeaderboardScreen = () => {
-  // This component would display the leaderboard
-  // based on the data loaded from the engagement store.
-  return (
-    <View flex center>
-      <Text text50b marginB-20>Leaderboard</Text>
-      <Text>(The leaderboard would be displayed here.)</Text>
-    </View>
-  );
+  return engagement.renderLeaderboardView("LeaderboardScreen");
 };
 
-// Wrap the component with the HOC to load engagement data
 export default withEngagementData(LeaderboardScreen);
