+++ client/src/learning/components/QuizView.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View, Text } from 'react-native-ui-lib';
+
+const QuizView = () => {
+  return (
+    <View flex center>
+      <Text text50b marginB-20>Quiz Screen</Text>
+      <Text>(Quiz content and logic would be displayed here.)</Text>
+    </View>
+  );
+};
+
+export default QuizView;
