+++ client/src/learning/screens/SessionSummaryScreen.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
-import { View, Text } from 'react-native-ui-lib';
+import { LearningComponent } from '..';
 import { withLearningData } from '../../common/hocs/withInitialData';
 
+const learning = new LearningComponent();
+
 const SessionSummaryScreen = () => {
-  // In a real implementation, this component would use hooks
-  // to access the session summary data from the Redux store.
-  return (
-    <View flex center>
-      <Text text50b marginB-20>Session Summary</Text>
-      <Text>(A summary of the user's performance would be displayed here.)</Text>
-    </View>
-  );
+  return learning.renderSummary("SessionSummaryScreen");
 };
 
-// Wrap the component with the HOC to load learning data
 export default withLearningData(SessionSummaryScreen);
