+++ client/src/mediator/screens/QuizScreen.tsx
@@ -15,7 +15,7 @@ const QuizScreen = ({ navigation }: any) => {
         <View row spread centerV paddingH-18 paddingT-10 paddingB-8 bg-white>
             <Button link iconSource={() => <Icon name="chevron-left" size={28} color="#222" />} onPress={() => navigation.goBack()} />
         </View>
-        {learning.renderQuizView("QuizScreen", handleNext)}
+        {learning.renderQuiz("QuizScreen", handleNext)}
     </View>
   );
 };
